T996:orkaudio:reporting: logging fixup on Windows
 - last T996 change (ef4b79bd685eb7885bb7e9a30dd3edc699b9d2b8)
   breaks Windows: timedeltasec is printed as a 32-bit
   number but is passed as a 64-bit number, screwing up the
   printing of the serialized tape message.

+++ orkbasecxx/Reporting.cpp
@@ -356,7 +356,7 @@ void ReportingThread::Run()
 
 				if( CONFIG.m_enableReporting)
 				{
-					FLOG_INFO(LOG.reporting,"[%s] timedeltasecs:%d sending: %s", m_tracker.ToString(), msgRef->Age(), msgAsSingleLineString);
+					FLOG_INFO(LOG.reporting,"[%s] timedeltasecs:%d sending: %s", m_tracker.ToString(), (int)(msgRef->Age()), msgAsSingleLineString);
 
 					//OrkHttpSingleLineClient c;
